+++ src/main/java/org/zalando/nakadi/controller/CursorOperationsController.java
@@ -4,13 +4,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.NakadiCursorLag;
@@ -18,7 +16,6 @@ import org.zalando.nakadi.domain.ShiftedNakadiCursor;
 import org.zalando.nakadi.exceptions.runtime.CursorConversionException;
 import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.InvalidCursorException;
-import org.zalando.nakadi.exceptions.runtime.InvalidCursorOperation;
 import org.zalando.nakadi.exceptions.runtime.NakadiBaseException;
 import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.NotFoundException;
@@ -32,9 +29,6 @@ import org.zalando.nakadi.view.Cursor;
 import org.zalando.nakadi.view.CursorDistance;
 import org.zalando.nakadi.view.CursorLag;
 import org.zalando.nakadi.view.ShiftedCursor;
-import org.zalando.problem.MoreStatus;
-import org.zalando.problem.Problem;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -43,7 +37,6 @@ import java.util.stream.Collectors;
 
 import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.ResponseEntity.status;
-import static org.zalando.problem.MoreStatus.UNPROCESSABLE_ENTITY;
 
 @RestController
 public class CursorOperationsController {
@@ -131,35 +124,6 @@ public class CursorOperationsController {
 
         return lagResult.stream().map(this::toCursorLag)
                 .collect(Collectors.toList());
-    }
-
-    @ExceptionHandler(InvalidCursorOperation.class)
-    public ResponseEntity<?> invalidCursorOperation(final InvalidCursorOperation e,
-                                                    final NativeWebRequest request) {
-        LOG.debug("User provided invalid cursor for operation. Reason: " + e.getReason(), e);
-        return Responses.create(Problem.valueOf(MoreStatus.UNPROCESSABLE_ENTITY,
-                clientErrorMessage(e.getReason())), request);
-    }
-
-    @ExceptionHandler(CursorConversionException.class)
-    public ResponseEntity<Problem> handleCursorConversionException(final CursorConversionException exception,
-                                                                   final NativeWebRequest request) {
-        LOG.error(exception.getMessage(), exception);
-        return Responses.create(UNPROCESSABLE_ENTITY, exception.getMessage(), request);
-    }
-
-    private String clientErrorMessage(final InvalidCursorOperation.Reason reason) {
-        switch (reason) {
-            case TIMELINE_NOT_FOUND: return "Timeline not found. It might happen in case the cursor refers to a " +
-                    "timeline that has already expired.";
-            case PARTITION_NOT_FOUND: return "Partition not found.";
-            case CURSOR_FORMAT_EXCEPTION: return "Сursor format is not supported.";
-            case CURSORS_WITH_DIFFERENT_PARTITION: return "Cursors with different partition. Pairs of cursors should " +
-                    "have matching partitions.";
-            default:
-                LOG.error("Unexpected invalid cursor operation reason " + reason);
-                throw new NakadiBaseException();
-        }
     }
 
     private CursorLag toCursorLag(final NakadiCursorLag nakadiCursorLag) {