+++ src/main/java/org/zalando/nakadi/controller/SubscriptionController.java
@@ -1,10 +1,7 @@
 package org.zalando.nakadi.controller;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -14,45 +11,31 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.nakadi.domain.ItemsWrapper;
 import org.zalando.nakadi.domain.SubscriptionEventTypeStats;
 import org.zalando.nakadi.exceptions.runtime.DbWriteOperationsBlockedException;
-import org.zalando.nakadi.exceptions.runtime.ErrorGettingCursorTimeLagException;
-import org.zalando.nakadi.exceptions.runtime.FeatureNotAvailableException;
 import org.zalando.nakadi.exceptions.runtime.InconsistentStateException;
 import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.InvalidLimitException;
-import org.zalando.nakadi.exceptions.runtime.NakadiBaseException;
 import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
-import org.zalando.nakadi.service.FeatureToggleService;
 import org.zalando.nakadi.service.subscription.SubscriptionService;
 import org.zalando.nakadi.service.subscription.SubscriptionService.StatsMode;
-import org.zalando.problem.Problem;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.annotation.Nullable;
 import java.util.Set;
 
-import static javax.ws.rs.core.Response.Status.NOT_IMPLEMENTED;
-import static javax.ws.rs.core.Response.Status.SERVICE_UNAVAILABLE;
 import static org.springframework.http.HttpStatus.NO_CONTENT;
 import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.ResponseEntity.status;
-import static org.zalando.problem.MoreStatus.UNPROCESSABLE_ENTITY;
 
 
 @RestController
 @RequestMapping(value = "/subscriptions")
 public class SubscriptionController {
 
-    private static final Logger LOG = LoggerFactory.getLogger(SubscriptionController.class);
-
-    private final FeatureToggleService featureToggleService;
     private final SubscriptionService subscriptionService;
 
     @Autowired
-    public SubscriptionController(final FeatureToggleService featureToggleService,
-                                  final SubscriptionService subscriptionService) {
-        this.featureToggleService = featureToggleService;
+    public SubscriptionController(final SubscriptionService subscriptionService) {
         this.subscriptionService = subscriptionService;
     }
 
@@ -94,30 +77,5 @@ public class SubscriptionController {
             NoSuchEventTypeException, NoSuchSubscriptionException, ServiceTemporarilyUnavailableException {
         final StatsMode statsMode = showTimeLag ? StatsMode.TIMELAG : StatsMode.NORMAL;
         return subscriptionService.getSubscriptionStat(subscriptionId, statsMode);
-    }
-
-    @ExceptionHandler(FeatureNotAvailableException.class)
-    public ResponseEntity<Problem> handleFeatureTurnedOff(final FeatureNotAvailableException ex,
-                                                          final NativeWebRequest request) {
-        LOG.debug(ex.getMessage(), ex);
-        return Responses.create(Problem.valueOf(NOT_IMPLEMENTED, ex.getMessage()), request);
-    }
-
-    @ExceptionHandler(ErrorGettingCursorTimeLagException.class)
-    public ResponseEntity<Problem> handleTimeLagException(final ErrorGettingCursorTimeLagException ex,
-                                                          final NativeWebRequest request) {
-        LOG.debug(ex.getMessage(), ex);
-        return Responses.create(Problem.valueOf(UNPROCESSABLE_ENTITY, ex.getMessage()), request);
-    }
-
-    @ExceptionHandler({InconsistentStateException.class, ServiceTemporarilyUnavailableException.class})
-    public ResponseEntity<Problem> handleServiceUnavailableResponses(final NakadiBaseException exception,
-                                                                     final NativeWebRequest request) {
-        LOG.debug(exception.getMessage(), exception);
-        return Responses.create(
-                Problem.valueOf(
-                        SERVICE_UNAVAILABLE,
-                        exception.getMessage()),
-                request);
     }
 }