+++ src/main/java/org/zalando/nakadi/exceptions/runtime/BlockedException.java
@@ -1,0 +1,8 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class BlockedException extends NakadiBaseException {
+
+    public BlockedException(final String message) {
+        super(message);
+    }
+}