+++ src/test/java/org/zalando/nakadi/controller/EventPublishingControllerTest.java
@@ -12,6 +12,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.zalando.nakadi.config.SecuritySettings;
+import org.zalando.nakadi.controller.advice.EventPublishingExceptionHandler;
+import org.zalando.nakadi.controller.advice.NakadiProblemExceptionHandler;
 import org.zalando.nakadi.domain.BatchItemResponse;
 import org.zalando.nakadi.domain.EventPublishResult;
 import org.zalando.nakadi.domain.EventPublishingStatus;
@@ -92,6 +94,7 @@ public class EventPublishingControllerTest {
         mockMvc = standaloneSetup(controller)
                 .setMessageConverters(new StringHttpMessageConverter(), TestUtils.JACKSON_2_HTTP_MESSAGE_CONVERTER)
                 .setCustomArgumentResolvers(new ClientResolver(settings, featureToggleService))
+                .setControllerAdvice(new NakadiProblemExceptionHandler(), new EventPublishingExceptionHandler())
                 .build();
     }
 
@@ -159,7 +162,7 @@ public class EventPublishingControllerTest {
     @Test
     public void whenEventTypeNotFoundThen404() throws Exception {
         Mockito
-                .doThrow(NoSuchEventTypeException.class)
+                .doThrow(new NoSuchEventTypeException("topic not found"))
                 .when(publisher)
                 .publish(any(String.class), eq(TOPIC));